revert remodeling to fit OBI
Rather than reviewing the existing surveillance activity structure, it was better to model the connection between HSO’s surveillance activity and OBI’s surveillance process. Considering the scope of the latter specifically as “biomedical investigations”, surveillance processes can be equivalent to specific types of surveillance activities, such as surveys and source attribution studies. A surveillance process can therefore be a specific type of surveillance activity, in which there is a specific investigation, with planning, study design, and drawing conclusions from data.